requirements.txt
fastapi==0.115.9
anyio>=4.8.0,<5.0.0
websockets>=13.0.0,<15.1.0

langchain==0.3.0
langchain-core>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
langchain-text-splitters>=0.3.0,<0.4.0
langchain-chroma==0.1.4
langchain-google-genai>=2.0.0,<3.0.0

google-genai==1.9.0
google-ai-generativelanguage>=0.6.18,<0.7.0

chromadb>=0.4.0,<0.6.0 
pandas
uvicorn
python-dotenv
